requirements.txt
# Core IA e LLM
langchain>=0.1.0
langchain-community>=0.0.10
langchain-groq>=0.1.0

# APIs e requests
requests>=2.31.0
groq>=0.4.1
huggingface>=0.10.0

# Interface web
streamlit>=1.29.0

# Processamento de dados
pandas>=2.1.0
numpy>=1.24.3

# Utilidades
python-dotenv>=1.0.0
pydantic>=2.5.0

# Análise de texto
nltk>=3.8.1 